git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/branches/SHORT_ComponentSCAControllers@18779 28e8926c-6b08-0410-baaa-805c5e19b8d6

+++ src/Tests/functionalTests/processbuilder/test/WindowsAndLinuxTester.java
@@ -1,3 +1,39 @@
+/*
+ * ################################################################
+ *
+ * ProActive Parallel Suite(TM): The Java(TM) library for
+ *    Parallel, Distributed, Multi-Core Computing for
+ *    Enterprise Grids & Clouds
+ *
+ * Copyright (C) 1997-2011 INRIA/University of
+ *                 Nice-Sophia Antipolis/ActiveEon
+ * Contact: proactive@ow2.org or contact@activeeon.com
+ *
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Affero General Public License
+ * as published by the Free Software Foundation; version 3 of
+ * the License.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this library; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307
+ * USA
+ *
+ * If needed, contact us to obtain a release under GPL Version 2 or 3
+ * or a different license than the AGPL.
+ *
+ *  Initial developer(s):               The ActiveEon Team
+ *                        http://www.activeeon.com/
+ *  Contributor(s):
+ *
+ * ################################################################
+ * $$ACTIVEEON_INITIAL_DEV$$
+ */
 package functionalTests.processbuilder.test;
 
 import static org.junit.Assert.assertFalse;
@@ -5,7 +41,6 @@ import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 
 import java.io.File;
-import java.io.FileReader;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.ArrayList;
@@ -22,6 +57,7 @@ import org.objectweb.proactive.extensions.processbuilder.OSProcessBuilder;
 import org.objectweb.proactive.extensions.processbuilder.OSRuntime;
 import org.objectweb.proactive.extensions.processbuilder.OSUser;
 import org.objectweb.proactive.extensions.processbuilder.PAOSProcessBuilderFactory;
+import org.objectweb.proactive.extensions.processbuilder.exception.FatalProcessBuilderException;
 import org.objectweb.proactive.extensions.processbuilder.exception.OSUserException;
 import org.objectweb.proactive.extensions.processbuilder.stream.LineReader;
 
@@ -30,10 +66,12 @@ import functionalTests.FunctionalTest;
 
 public class WindowsAndLinuxTester extends FunctionalTest {
     final static boolean isLinux = System.getProperty("os.name").toLowerCase().startsWith("linux");
+    final static boolean isWindows = System.getProperty("os.name").toLowerCase().contains("windows");
 
     // Modify these to fit your environment
-    OSUser user;
-    OSUser userWPass;
+    OSUser userPasswd;
+    OSUser userSsh;
+
     String tempPath;
 
     static OSRuntime osRuntime;
@@ -46,14 +84,16 @@ public class WindowsAndLinuxTester extends FunctionalTest {
     // ------------------------------------ 
     @Before
     public void shouldRun() throws ProActiveException {
-
         String suser = System.getenv("OSPB_TEST_USER");
         Assume.assumeNotNull(suser, "process builder not tested because OSPB_TEST_USER is not set");
-        user = new OSUser(System.getenv("OSPB_TEST_USER"));
 
         String pass = System.getenv("OSPB_TEST_PASS");
         Assume.assumeNotNull(pass, "process builder not tested because OSPB_TEST_PASS is not set");
-        userWPass = new OSUser(suser, pass);
+        userPasswd = new OSUser(suser, pass);
+
+        //        String pass = System.getenv("OSPB_TEST_SSHKEY");
+        //        Assume.assumeNotNull(pass, "process builder not tested because OSPB_TEST_ is not set");
+        //        userPasswd = new OSUser(suser, null);
 
         tempPath = System.getenv("OSPB_TEST_TEMP");
         Assume.assumeNotNull(tempPath, "process builder not tested because OSPB_TEST_TEMP is not set");
@@ -87,7 +127,6 @@ public class WindowsAndLinuxTester extends FunctionalTest {
         } catch (Exception e) {
             assertNull(e.getMessage(), e);
         }
-
     }
 
     @Test
@@ -101,14 +140,16 @@ public class WindowsAndLinuxTester extends FunctionalTest {
         String[] expectedOut = { "111" };
         String[] expectedErr = { "222" };
 
-        try {
-            runAndMatch(cmd, user, dir, env, expectedOut, expectedErr, 0);
-        } catch (Exception e) {
-            assertNull(e.getMessage(), e);
+        if (!isWindows) {
+            try {
+                runAndMatch(cmd, userPasswd, dir, env, expectedOut, expectedErr, 0);
+            } catch (Exception e) {
+                assertNull(e.getMessage(), e);
+            }
         }
 
         try {
-            runAndMatch(cmd, userWPass, dir, env, expectedOut, expectedErr, 0);
+            runAndMatch(cmd, userPasswd, dir, env, expectedOut, expectedErr, 0);
         } catch (Exception e) {
             assertNull(e.getMessage(), e);
         }
@@ -123,30 +164,40 @@ public class WindowsAndLinuxTester extends FunctionalTest {
         File dir = null;
         HashMap<String, String> env = null;
 
-        String[] expectedOut = { user.getUserName() };
+        String[] expectedOut = { userPasswd.getUserName() };
+        if (isWindows) {
+            expectedOut[0] = System.getenv().get("USERDOMAIN") + "\\" + userPasswd.getUserName();
+        }
+
         String[] expectedErr = {};
 
-        try {
-            runAndMatch(cmd, user, dir, env, expectedOut, expectedErr, 0);
-        } catch (Exception e) {
-            assertNull(e.getMessage(), e);
-        }
-        try {
-            runAndMatch(cmd, userWPass, dir, env, expectedOut, expectedErr, 0);
-        } catch (Exception e) {
-            assertNull(e.getMessage(), e);
+        if (!isWindows) {
+            try {
+                runAndMatch(cmd, userPasswd, dir, env, expectedOut, expectedErr, 0);
+            } catch (Exception e) {
+                assertNull(e.getMessage(), e);
+            }
         }
 
+        try {
+            runAndMatch(cmd, userPasswd, dir, env, expectedOut, expectedErr, 0);
+        } catch (Exception e) {
+            assertNull(e.getMessage(), e);
+        }
     }
 
     @Test
     public void checkCanExecuteAsUser() throws Exception {
-        OSProcessBuilder ospb = new PAOSProcessBuilderFactory().getBuilder();
-        assertTrue(ospb.canExecuteAsUser(user));
-        assertTrue(ospb.canExecuteAsUser(userWPass));
-        assertFalse(ospb.canExecuteAsUser(new OSUser(user.getUserName(), "jibberish")));
-        assertFalse(ospb.canExecuteAsUser(new OSUser("jibberish")));
-
+        OSProcessBuilder ospbUser = new PAOSProcessBuilderFactory().getBuilder(userPasswd);
+        OSProcessBuilder ospbUserWPass = new PAOSProcessBuilderFactory().getBuilder(userPasswd);
+        if (!isWindows) {
+            assertTrue(ospbUser.canExecuteAsUser(userPasswd));
+        }
+        assertTrue(ospbUserWPass.canExecuteAsUser(userPasswd));
+        if (!isWindows) {
+            assertFalse(ospbUser.canExecuteAsUser(new OSUser(userPasswd.getUserName(), "jibberish")));
+            assertFalse(ospbUser.canExecuteAsUser(new OSUser("jibberish")));
+        }
     }
 
     @Test
@@ -202,18 +253,18 @@ public class WindowsAndLinuxTester extends FunctionalTest {
         String[] expectedOut = {};
         String[] expectedErr = {};
 
-        try {
-            runAndMatch(cmd, user, dir, env, expectedOut, expectedErr, null);
-
-        } catch (IOException e) {
-            // expected
-        } catch (Exception e) {
-            assertNull(e.getMessage(), e);
+        if (!isWindows) {
+            try {
+                runAndMatch(cmd, userPasswd, dir, env, expectedOut, expectedErr, null);
+            } catch (IOException e) {
+                // expected
+            } catch (Exception e) {
+                assertNull(e.getMessage(), e);
+            }
         }
 
         try {
-            runAndMatch(cmd, userWPass, dir, env, expectedOut, expectedErr, null);
-
+            runAndMatch(cmd, userPasswd, dir, env, expectedOut, expectedErr, null);
         } catch (IOException e) {
             // expected
         } catch (Exception e) {
@@ -227,27 +278,28 @@ public class WindowsAndLinuxTester extends FunctionalTest {
         //create reference value
         Process p = Runtime.getRuntime().exec(cmd);
         expectedErr = getError(p);
-        if (!isLinux) {
-            //windows will output a shorter error message if the command is 
-            //started from inside an other command... at east it seems so
-            String err[] = new String[1];
-            err[0] = expectedErr[0];
-            expectedErr = err;
-        }
+        //        if (!isLinux) {
+        //            //windows will output a shorter error message if the command is 
+        //            //started from inside an other command... at east it seems so
+        //            String err[] = new String[1];
+        //            err[0] = expectedErr[0];
+        //            expectedErr = err;
+        //        }
         expectedOut = getOutput(p);
         p.waitFor();
-        try {
-            runAndMatch(cmd, user, dir, env, expectedOut, expectedErr, null);
 
-        } catch (IOException e) {
-            // expected
-        } catch (Exception e) {
-            assertNull(e.getMessage(), e);
+        if (!isWindows) {
+            try {
+                runAndMatch(cmd, userPasswd, dir, env, expectedOut, expectedErr, null);
+            } catch (IOException e) {
+                // expected
+            } catch (Exception e) {
+                assertNull(e.getMessage(), e);
+            }
         }
 
         try {
-            runAndMatch(cmd, userWPass, dir, env, expectedOut, expectedErr, null);
-
+            runAndMatch(cmd, userPasswd, dir, env, expectedOut, expectedErr, null);
         } catch (IOException e) {
             // expected
         } catch (Exception e) {
@@ -289,14 +341,16 @@ public class WindowsAndLinuxTester extends FunctionalTest {
         String[] expectedErr = {};
         Integer exitValue = 1;
 
-        try {
-            runAndMatch(cmd, user, dir, env, expectedOut, expectedErr, exitValue);
-        } catch (Exception e) {
-            assertNull(e.getMessage(), e);
+        if (!isWindows) {
+            try {
+                runAndMatch(cmd, userPasswd, dir, env, expectedOut, expectedErr, exitValue);
+            } catch (Exception e) {
+                assertNull(e.getMessage(), e);
+            }
         }
 
         try {
-            runAndMatch(cmd, userWPass, dir, env, expectedOut, expectedErr, exitValue);
+            runAndMatch(cmd, userPasswd, dir, env, expectedOut, expectedErr, exitValue);
         } catch (Exception e) {
             assertNull(e.getMessage(), e);
         }
@@ -306,14 +360,16 @@ public class WindowsAndLinuxTester extends FunctionalTest {
         cmd = (isLinux) ? lc2 : wc2;
         exitValue = 111;
 
-        try {
-            runAndMatch(cmd, user, dir, env, expectedOut, expectedErr, exitValue);
-        } catch (Exception e) {
-            assertNull(e.getMessage(), e);
+        if (!isWindows) {
+            try {
+                runAndMatch(cmd, userPasswd, dir, env, expectedOut, expectedErr, exitValue);
+            } catch (Exception e) {
+                assertNull(e.getMessage(), e);
+            }
         }
 
         try {
-            runAndMatch(cmd, userWPass, dir, env, expectedOut, expectedErr, exitValue);
+            runAndMatch(cmd, userPasswd, dir, env, expectedOut, expectedErr, exitValue);
         } catch (Exception e) {
             assertNull(e.getMessage(), e);
         }
@@ -323,14 +379,16 @@ public class WindowsAndLinuxTester extends FunctionalTest {
         cmd = (isLinux) ? lc3 : wc3;
         exitValue = (isLinux) ? 250 : -1;
 
-        try {
-            runAndMatch(cmd, user, dir, env, expectedOut, expectedErr, exitValue);
-        } catch (Exception e) {
-            assertNull(e.getMessage(), e);
+        if (!isWindows) {
+            try {
+                runAndMatch(cmd, userPasswd, dir, env, expectedOut, expectedErr, exitValue);
+            } catch (Exception e) {
+                assertNull(e.getMessage(), e);
+            }
         }
 
         try {
-            runAndMatch(cmd, userWPass, dir, env, expectedOut, expectedErr, exitValue);
+            runAndMatch(cmd, userPasswd, dir, env, expectedOut, expectedErr, exitValue);
         } catch (Exception e) {
             assertNull(e.getMessage(), e);
         }
@@ -348,15 +406,22 @@ public class WindowsAndLinuxTester extends FunctionalTest {
         String[] expectedOut = { unicode };
         String[] expectedErr = {};
         Integer exitValue = 0;
-        try {
-            runAndMatch(cmd, user, dir, env, expectedOut, expectedErr, exitValue);
-        } catch (Exception e) {
-            assertNull(e.getMessage(), e);
+
+        if (!isWindows) {
+            try {
+                runAndMatch(cmd, userPasswd, dir, env, expectedOut, expectedErr, exitValue);
+            } catch (Exception e) {
+                assertNull(e.getMessage(), e);
+            }
         }
-        try {
-            runAndMatch(cmd, userWPass, dir, env, expectedOut, expectedErr, exitValue);
-        } catch (Exception e) {
-            assertNull(e.getMessage(), e);
+
+        // the default behavior (with Runtime.exec) is not accepted by this test
+        if (!isWindows) {
+            try {
+                runAndMatch(cmd, userPasswd, dir, env, expectedOut, expectedErr, exitValue);
+            } catch (Exception e) {
+                assertNull(e.getMessage(), e);
+            }
         }
 
     }
@@ -372,13 +437,17 @@ public class WindowsAndLinuxTester extends FunctionalTest {
         String[] expectedOut = { dir.getCanonicalPath() };
         String[] expectedErr = {};
         Integer exitValue = 0;
-        try {
-            runAndMatch(cmd, user, dir, env, expectedOut, expectedErr, exitValue);
-        } catch (Exception e) {
-            assertNull(e.getMessage(), e);
+
+        if (!isWindows) {
+            try {
+                runAndMatch(cmd, userPasswd, dir, env, expectedOut, expectedErr, exitValue);
+            } catch (Exception e) {
+                assertNull(e.getMessage(), e);
+            }
         }
+
         try {
-            runAndMatch(cmd, userWPass, dir, env, expectedOut, expectedErr, exitValue);
+            runAndMatch(cmd, userPasswd, dir, env, expectedOut, expectedErr, exitValue);
         } catch (Exception e) {
             assertNull(e.getMessage(), e);
         }
@@ -387,24 +456,23 @@ public class WindowsAndLinuxTester extends FunctionalTest {
         assertTrue(dir.mkdir());
         expectedOut[0] = dir.getCanonicalPath();
 
-        try {
-            runAndMatch(cmd, user, dir, env, expectedOut, expectedErr, exitValue);
-
-        } catch (Exception e) {
-            dir.delete();
-            assertNull(e.getMessage(), e);
+        if (!isWindows) {
+            try {
+                runAndMatch(cmd, userPasswd, dir, env, expectedOut, expectedErr, exitValue);
+            } catch (Exception e) {
+                dir.delete();
+                assertNull(e.getMessage(), e);
+            }
         }
 
         try {
-            runAndMatch(cmd, userWPass, dir, env, expectedOut, expectedErr, exitValue);
-
+            runAndMatch(cmd, userPasswd, dir, env, expectedOut, expectedErr, exitValue);
         } catch (Exception e) {
             dir.delete();
             assertNull(e.getMessage(), e);
         }
 
         dir.delete();
-
     }
 
     @Test
@@ -421,17 +489,33 @@ public class WindowsAndLinuxTester extends FunctionalTest {
         String[] expectedErr = {};
         Integer exitValue = 0;
 
-        try {
-            runAndMatch(cmd, user, dir, env, expectedOut, expectedErr, exitValue);
-        } catch (Exception e) {
-            assertNull(e.getMessage(), e);
-        }
-        try {
-            runAndMatch(cmd, userWPass, dir, env, expectedOut, expectedErr, exitValue);
-        } catch (Exception e) {
-            assertNull(e.getMessage(), e);
+        if (!isWindows) {
+            try {
+                runAndMatch(cmd, userPasswd, dir, env, expectedOut, expectedErr, exitValue);
+            } catch (Exception e) {
+                assertNull(e.getMessage(), e);
+            }
         }
 
+        // On Windows the behaviour is not defined yet
+        // if the forked process as user inherits the environment from its creator it
+        // can represent a security issue ..
+        if (!isWindows) {
+            try {
+                runAndMatch(cmd, userPasswd, dir, env, expectedOut, expectedErr, exitValue);
+            } catch (Exception e) {
+                assertNull(e.getMessage(), e);
+            }
+        }
+
+    }
+
+    @Test
+    public void testDestroy() throws IOException, OSUserException, FatalProcessBuilderException {
+
+        Process p = osRuntime.exec(this.userPasswd, new String[] { "/bin/sleep", "5555" }, null,
+                new File("."));
+        p.destroy();
     }
 
     public static void runAndMatch(String[] cmd, OSUser user, File dir, Map<String, String> env,
@@ -445,22 +529,23 @@ public class WindowsAndLinuxTester extends FunctionalTest {
             //check output
             for (int i = 0; i < out.length; i++) {
                 if (!out[i].equals(expectedOut[i]))
-                    errors += ("Outputs don't match at line " + i + "\n I have\"" + out[i] +
+                    errors += ("Outputs don't match at line " + i + "\n I have \"" + out[i] +
                         "\" instead of \"" + expectedOut[i] + "\"\n");
             }
             //check error
             for (int i = 0; i < err.length; i++) {
                 if (!err[i].equals(expectedError[i]))
-                    errors += ("Errors don't match at line " + i + "\n I have\"" + err[i] +
+                    errors += ("Errors don't match at line " + i + "\n I have \"" + err[i] +
                         "\" instead of \"" + expectedError[i] + "\"\n");
             }
         } else {
             errors += ("Output lengths does not match!\nReal Out:" + out.length + " Exp. Out:" +
                 expectedOut.length + "\nReal Err: " + err.length + " Exp. Err:" + expectedError.length);
         }
-        int eval = p.waitFor();
+        p.waitFor();
+        int eval = p.exitValue();
         if (expectedRet != null && !expectedRet.equals(eval))
-            errors += ("Exit valuse is " + eval + " instead of " + expectedRet);
+            errors += ("Exit value is " + eval + " instead of " + expectedRet);
 
         if (errors.length() > 0) {
             throw new Exception(errors);
@@ -472,11 +557,9 @@ public class WindowsAndLinuxTester extends FunctionalTest {
         InputStream inputstream = p.getInputStream();
         LineReader bufferedreader = new LineReader(inputstream);
         String line;
-
         while ((line = bufferedreader.readLine()) != null) {
             list.add(line);
         }
-
         return list.toArray(new String[0]);
     }
 
@@ -489,8 +572,6 @@ public class WindowsAndLinuxTester extends FunctionalTest {
         while ((line = bufferedreader.readLine()) != null) {
             list.add(line);
         }
-
         return list.toArray(new String[0]);
     }
-
 }
